tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/constants/001.html

<!doctype html>
<title>WebSockets: getting constants on constructor</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
var constants = ['CONNECTING', 'OPEN', 'CLOSING', 'CLOSED'];
for (var i = 0; i < constants.length; ++i) {
  test(function(){
    var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
    assert_equals(WebSocket[constants[i]], i, 'WebSocket.'+constants[i]);
  }, "Constants on constructors " + constants[i]);
};
</script>
